Don't throttle ethernet when on AC Power
Can't see a good reason why one should throttle the ethernet speed when 
running on AC Power.
As for when running on Battery, we still have a use case where the user
might still want to run at full speed

Please let me know if there are any comments on this change.