Run pidof with the -x Script Mode switch. Thanks Matthijs Kooijman
while debugging another issue, I noticed that the check for running lm-polling
daemons is not working.
/usr/share/laptop-mode-tools/modules/battery-level-polling uses
"pidof lm-polling-daemon" to see if the daemon is already running, but this
does not work. The actual process is /bin/sh, which is running the
lm-polling-daemon script. To make this work properly, pidof supports a -x
option: 

       -x     Scripts too - this causes the program to also return process
              id’s of shells running the named scripts.

When passing -x, laptop-mode stops spawning new on every restart (and probably
in other cases as well). It seems that the killall line used does work
properly.