Disconnect descriptors when backgrounding the shell script. Thanks Matthijs Kooijman
> >   /usr/share/laptop-mode-tools/module-helpers/lm-polling-daemon \
> >   < /dev/null > /dev/null 2> /dev/null &
> 
> Since we are disconnecting it from all the descriptors, do we need the "&" ?
Yes, since otherwise the script would wait right there until the polling
daemon terminates (ie, forever).

> Also, this would have not been correct if we were using echo to print 
> messages. But since we use a logger now, it won't affect us.
Right. However, the only way for the polling daemon to use stdout would be to
not capture laptop_mode's output in the init script. You can't do both.
However, using a logger is the proper way anyway, since it would be weird (and
not all too useful) for a daemon to output to the console on which it happened
to be started (which could be no console at all, if it was started from an
ACPI event handler, for example!).