Write to /dev/dsp in a sub-shell
We do this to be able to not clutted our display.

========================
Thanks to Preben Guldberg.

How about this one:

        (exec 2>/dev/null; echo 1 >/dev/dsp)

Afaik it ought to work without the subshell, but on a random linux
host in the lab here bash was not too fond of that (zsh liked it, btw).

Preben

-----Original Message-----
From: Sarraf, Ritesh 
Sent: Monday, August 31, 2009 08:30
To: c-dl-linux
Subject: Redirecting echo failures

Hi,

champaran:~# echo 1 >/dev/dsp
-su: /dev/dsp: Device or resource busy


I'd like to redirect this failure to /dev/null.
This failure is non-fatal and I just don't want to see it. Any suggestions ?

Background:
On power switch, for changes to low level power settings in audio drivers 
(ac97 and intel-hda), the timer changes only kick in after the device is 
closed/opened. The above command does that.
The failure occurs if the audio device is busy. (In cases where the device is 
busy, worst is the timer changes will not be effective).


Ritesh