Introduce individual debug mode for each module
Debugging LMT can be a PITA especially if you enable the global DEBUG=1 flag.
You can easily lose track of where to look for.
When you need to debug the behavior of a particular module only, it makes sense
to enable debug mode only for that particular module. This commit helps with that.

With this feature in place, I hope more users can contribute to LMT. It will be
easier for users to monitor if LMT is properly behaving on their machines or not.

Brief testing is done and things work as expected. Thorough testing will be done before the final release