pass arg to the calling function
On Fri, 18 Feb 2011 17:45:18 +0530
Ritesh Raj Sarraf <rrs@researchut.com> wrote:

> > Can you elaborate what you mean by "discards all the arguments given
> > to it" ?
> > How are you calling lmt?
I'm talking about /usr/sbin/laptop_mode file, which is a shell script.
Man page says that is should accept "auto" and "status" as a command
and an optional "force" argument.
I've looked through this script and have found out that it ignores any
arguments it gets. It tries to read positional parameters in function
called "lmt_main_function", but what it reads are the parameters of
function, not of the script. And as the script doesn't pass its
parameters to this function, they are effectively ignored. My patch
changes this, so that arguments are recognized.

I hope this explains the issue sufficiently.