Support execution of complex commands
Thanks to Changaco for the patch.

It seems that the exec-commands module cannot execute "complex" commands, for example containing an output redirection ">". I'm sending you a patch that fixes this for me.

I was trying to see if I could use exec-commands to shut down the second core of my CPU with "echo 0 > /sys/devices/system/cpu/cpu1/online". Do you think that is useful ?