USB auto-suspend whitelist
Thanks: Simon Que

It rearranges the
blacklist functions and introduces an alternative whitelist system
that is "opt-in" rather than "opt-out" based on devices and device
types listed.  One can switch between using the blacklists and the
whitelists with the AUTOSUSPEND_USE_WHITELIST variable in
usb-autosuspend.conf.

With whitelists, we can have a setup where usb-autosuspend will not
run for any devices that are not known to be integral to the system.
With blacklists, this would have been tedious as every new device
would have to be blacklisted.