usb autosuspend black-/whitelist arg in quotes
Thanks: Simon Que

Attached is a patch to fix a bug in my previous usb-autosuspend
whitelist patch.  When the blacklists and whitelists are passed into
listed_by_id() and listed_by_type(), they need to be in quotation marks
in order to be treated as a single argument ($2) by those functions.
Otherwise they will generate one argument per list item.