Append to stdout/stderr to avoid truncating file logs
Thanks: Bjørn Mork
Closes: #668611

I noticed that my  /var/log/pm-suspend.log always would start with the
line

  Module /etc/laptop-mode/modules/* is not executable or is to be
skipped.

and missing a lot of expected messages before this line.  Found that the
reason was that I had VERBOSE_OUTPUT enabled in
/etc/laptop-mode/laptop-mode.conf. This makes laptop_mode write messages
to stderr by redirecting to /dev/fd/2

Problem is that if already redirected, like when running from the
/usr/lib/pm-utils/sleep.d/01laptop-mode hook, then this will truncate
the file stderr is redirected to.  The attached patch fix this by
appending to /dev/fd/2 instead.

This fixes my problem with a truncated /var/log/pm-suspend.log