Fix path for enable / disable ethernet device
+ [ x1 = x1 ]
+ [ -x /usr/bin/logger -a VERBOSE != STATUS ]
+ [ VERBOSE = MSG ]
+ [ VERBOSE = ERR ]
+ [ VERBOSE = VERBOSE ]
+ [ x0 = x1 ]
+ [ VERBOSE = VERBOSE ]
+ [ 1 = 0 ]
+ [ -z /sys/devices/pci0000:00/0000:00:19.0 ]
+ [ -f /sys/devices/pci0000:00/0000:00:19.0/enabled ]
+ [ x1 = x1 ]
+ [ -f /sys/devices/pci0000:00/0000:00:19.0/enabled ]
+ echo 0
sh: echo: I/O error
+ log VERBOSE ethernet: Disabling ethernet device eth0
+ [ x1 = x1 ]
+ [ -x /usr/bin/logger -a VERBOSE != STATUS ]
+ [ VERBOSE = MSG ]
+ [ VERBOSE = ERR ]

For some reason, I/O error is seen when writing to the file.
This seems to be a cosmetic timing issue because irrespective of the
error, the values are written to the file, and changes are effective.

It could very well be a timing issue because the ethernet device is a
PCI device too, which is also handle in Runtime PM