Detect ethernet carrier, early and relibaly
Thanks: http://www.dmnc.net

Hi, I'm using laptop-mode-tools in default configuration on Arch linux,
I found a problem with DISABLE_ETHERNET_ON_BATTERY option (set to 1),
regarding to the comment in the config file it may disable the interface
only if cable is not plugged in but it's disabling it even with the
active connection.

Reason for this behavior is *xxx_THROTTLE_ETHERNET" option. Even if all
3 speed throttling options are disabled, the script is always touching
it via ethtool: ethernet:117

:warning: And now, every ethtool speed manipulation resolves in
temporary down-link (in my case it takes 5s until kernel brings back the
link up) and due to this, ip gets NO-CARRIER.

Because script is checking the link status just after the speed
manipulation (ethernet:129), I've separated the carrier detection
higher.

Result: We know the carrier status from earlier and we're just working
with $carrier and not doing late detection which can be affected by
ethtool speed manipulations (if link has been lost for couple of
seconds).

Minor changes:

    setting the speed: in my case I'm getting "Cannot advertise speed
xxx" error, I found proper syntax with autoneg [off|on], i.e.: ethtool
-s eth0 speed 100 autoneg off
    restoring the speed: dtto, ethtool -s eth0 autoneg on

Cheers