Narrow down our power_supply subsystem event check condition
So the thing is... power_supply subsystems seems to be emitting events
on every battery change: things like charge of the battery. While
charging and while draining.

So if we honor power_supply as a subsystem and look for its events, it
can be way too many, like a loop.

What we care if only plug/unplug. So narrow it down.

This bug/behaviro ate my weekend. Details have been logged into the
following bug report:

https://github.com/rickysarraf/laptop-mode-tools/issues/45

On that bug report, there are other (systemd) issues also But this one,
now with this commit, should be fixed