Fine tune the .service file
Quoting the systemd.service manpage:

When Type is not oneshot, only one command may and must be given. When Type=oneshot is used, zero or more commands may be
specified. This can be specified by providing multiple command lines in the same directive, or alternatively, this directive may be specified more than
once with the same effect. If the empty string is assigned to this option, the list of commands to start is reset, prior assignments of this option
will have no effect. If no ExecStart= is specified, then the service must have RemainAfterExit=yes set.

For each of the specified commands, the first argument must be an absolute path to an executable. Optionally, if this file name is
prefixed with "@", the second token will be passed as "argv[0]" to the executed process, followed by the further arguments specified. If the absolute
filename is prefixed with "-", an exit code of the command normally considered a failure (i.e. non-zero exit status or abnormal exit due to signal) is
ignored and considered success. If both "-" and "@" are used, they can appear in either order.