Drop out reload
Problem is, that at some times, the script is started early during boot
or by udev, and not all devices are initialized in power save mode.

So when calling just "auto", LMT is not scanning the remaining list of
devices.

By dropping the "reload" target, we let systemd handle reload the
traditional way, i.e. stop n' start