Wait for all forked modules to complete
Thanks: Tomas Janousek

Background: wait was removed in 2012 because because of it, the lock was
being inherited by the polling daemon, or atleast that is what I thought
then. Now with systemd the default, and our new polling daemon is timer
based, there is no need of a battery polling daemon.