Add new module: cputhotplug
This module, if activated, will offline all your non-boot CPU (and
logical CPUs, like cores etc), and may result in some power savings.

I don't have profiling data to back, but in my opinion, it makes sense
to have this feature. It should not be enabled by default for all
ON_BATTERY cases, but there are cases where we should use it.
Like: the new hybrid machine, when run in Table mode, can be made to
shutdown all additional CPUs. As a matter of fact, some of the device
vendors do it right in the firmware (Lenovo Yoga 3).

So having it in userspace, and that too as an additional module is not
going to bite. :-) It just will give more choice to the user