Restore to MAX speed when back to AC Power
This will ensure that if the user is using throttling, then we throttle
the speed as requested by the user. THis still has the caveat that we
don't offer an ON_AC speed. But would anyone really want to run below
MAX_SPEED when on AC ? I don't think so. The only case where they'd want
to do that is, when there is a driver bug.

So, unless there are user reports, this should be the sane default

Now, to the other importat scenario. If the user does NOT enable
throttling, do not touch speed settings at all. Network drivers are
such, that every time you touch the speed setting, they re-negotiate.
And in that process most drivers tend to re-initialize the link