Update default list of USB type devices
The current device listing was wrong; the DRIVER field does not seem to
have usbmouse or usbkbd as driver type. I don't know what led me to that
assumption.

I've verified now, myself, with a USB Keyboard and Mouse, and the driver
is not usbkbd or usbmouse

IMPORTANT: Packagers are advised to cherry pick this commit on top of the
1.69 release. This commit seems to now completely fix the autosuspend
issue