Fix defaults for Intel P State Driver
Thanks: George Caswell and Matthew Gabeler-Lee
Closes: #814578

The LMT P State module's history shows the ONBATT default value of 50. I
have no history on why that was chosen. But reading up the docs, and
looking at  the feedback from reporters (thanks), this seems to be
wrong.

Intel P State driver is self sufficient in scaling in/out the frequency
and voltage based on the state of the machine. There isn't a need to do
much from sysfs, for the default cases.

Reference:
https://software.intel.com/en-us/articles/power-management-states-p-states-c-states-and-package-c-states#_Toc383778909