Added delay to battery status check
This check always jumps the gun on my hardware - ThinkPad x230

Without this delay, LMT reports a broken battery, every time the machine is unplugged.  
 
But that's only because the battery status is a little slower to update than laptop-mode is.  With the delay, it correctly reads the status, every time.