Prefer the current capacity of the battery over the original design capacity when calculating charge levels
The design capacity of the battery is only accurate for the first few days or weeks of a battery's life.  

The current capacity should **always** be preferred over the design capacity in these calculations.  This change causes laptops to actually hibernate AT the percentage you set, rather than above it.

I have applied this change to my /usr/sbin/laptop_mode, and it is working PERFECTLY now.  It hibernates exactly when it should, and reports an accurate, current value for the max charge in logs.