Add btrfs to list of filesystems for which we can set commit interval
Since commit ee5d8eeee254c24029c04a0214aa59bc20996c84 in 2012 an ability
to set a commit interval was added to btrfs (in kernel 3.13, released in
early 2014) so let's make use of it.

Without this change a default btrfs commit interval of 30 seconds results
in HDD spinning up very often when working on battery power.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>