Mark ethernet module disabled by default
Because these days, Linux gives unique device id to LOM devices.
Instead of making a guess, let's leave it to the user to specify it

Disable ethernet module by default because many ethernet devices were
already reported to be malfunctioning in throttling mode, which
resulted in us disabling throttling mode for it all these years.

So, disable this module by default and leave it to the user to enable it