Try harder to determine the cpu frequencies
On some user systems reported, the scaling_available_frequencies may not
be available through sysfs at all. For example, in bug report:
https://github.com/rickysarraf/laptop-mode-tools/issues/146, it has been
reported such.

```
~ $ cpupower frequency-info
analyzing CPU 0:
  driver: intel_pstate
  CPUs which run at the same hardware frequency: 0
  CPUs which need to have their frequency coordinated by software: 0
  maximum transition latency:  Cannot determine or is not supported.
  hardware limits: 800 MHz - 3.60 GHz
  available cpufreq governors: performance powersave
  current policy: frequency should be within 800 MHz and 3.60 GHz.
                  The governor "powersave" may decide which speed to use
                  within this range.
  current CPU frequency: Unable to call hardware
  current CPU frequency: 2.25 GHz (asserted by call to kernel)
  boost state support:
    Supported: yes
    Active: yes
```

This is on a machine with the `intel_pstate` driver.

Since, there's not much LMT can do than try harder, in cases where the
scaling_available_frequencies interface is not available, we fallback to
using half of the max frequency reported by the CPU.

Thanks: Thanatermesis
Closes: https://github.com/rickysarraf/laptop-mode-tools/issues/146