laptop_mode: fix restoring filesystem commit interval
grep without "-E" option uses basic regular expressions, where "+"
repetition operator needs to be backslashed as "\+", but
replace_numeric_mount_option() function used just a plain "+".

This resulted in this function being unable to find a commit interval
option in the saved mount options and so substituting it with the default
"commit=0" instead.

Due to this mistake any non-default commit interval would get reset to the
filesystem-specific default value (as provided by "commit=0") when laptop
mode was deactivated (like when the AC supply was plugged).