Use newer tracing interface instead of block_dump
Newer Linux releases, starting with 5.14+, dropped the legacy
/proc/sys/vm/block_dump interface.

Users are encouraged to use the tracing infrastructure instead, which is
what this change brings.

Note: block_dump enabling was a global knob for the entire block layer.
On the other hand, the tracing mechanism is on a per block device level.
As such, in this proposed change, we add small checks to extract the
list of block devices

Closes: https://github.com/rickysarraf/laptop-mode-tools/issues/202